refactor(next.config): convert to esm
